refactor: Code improvements for example 9
